mypt-cushomemodel-auth-api/requirements.txt
asgiref==3.5.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
cron-schedule==1.1
croniter==1.3.8
cryptography==40.0.2
Deprecated==1.2.13
deprecation==2.1.0
Django==4.1.3
django-cors-headers==3.13.0
django-enumfield==3.1
django-filter==22.1
django-ratelimit==4.0.0
django-redis==5.2.0
djangorestframework==3.14.0
emoji==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.0.4
h11==0.14.0
idna==3.4
img2pdf==0.4.4
kafka-python==2.0.2
lxml==4.9.2
Markdown==3.4.1
mysqlclient==2.1.1
numpy==1.23.4
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pikepdf==7.1.2
Pillow==9.3.0
pillow-heif==0.10.1
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
pyshorteners==1.0.1
PySocks==1.7.1
python-cron==1.0.1
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.0
pytimeparse==1.1.8
pytz==2022.6
redis==4.3.4
requests==2.28.1
rsa==4.9
schedule==1.1.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.6
urllib3==1.26.12
uuid==1.30
webdriver-manager==3.8.5
wrapt==1.14.1
wsproto==1.2.0
